Moved cleanup of old data files to cleanup folder
